requirements.txt
pillow
pathlib
tabulate